Support ref/definitions in JSON Schema
