+++ typesystem/__init__.py
@@ -1,4 +1,3 @@
-from typesystem.definitions import SchemaDefinitions
 from typesystem.fields import (
     Array,
     Boolean,
@@ -10,7 +9,6 @@ from typesystem.fields import (
     Float,
     Integer,
     Object,
-    Reference,
     String,
     Text,
     Time,
@@ -18,7 +16,7 @@ from typesystem.fields import (
 )
 from typesystem.forms import Jinja2Forms
 from typesystem.json_schema import from_json_schema, to_json_schema
-from typesystem.schemas import Schema
+from typesystem.schemas import Reference, Schema, SchemaDefinitions
 from typesystem.tokenize.tokenize_json import validate_json
 from typesystem.tokenize.tokenize_yaml import validate_yaml
