+++ typesystem/schemas.py
@@ -1,10 +1,48 @@
 import typing
 from abc import ABCMeta
-from collections.abc import Mapping
+from collections.abc import Mapping, MutableMapping
 
 from typesystem.base import ValidationError, ValidationResult
-from typesystem.definitions import SchemaDefinitions, set_definitions
-from typesystem.fields import Field, Object
+from typesystem.fields import Array, Field, Object
+
+
+class SchemaDefinitions(MutableMapping):
+    def __init__(self, *args: typing.Any, **kwargs: typing.Any) -> None:
+        self._definitions = dict(*args, **kwargs)  # type: dict
+
+    def __getitem__(self, key: typing.Any) -> typing.Any:
+        return self._definitions[key]
+
+    def __iter__(self) -> typing.Iterator[typing.Any]:
+        return iter(self._definitions)
+
+    def __len__(self) -> int:
+        return len(self._definitions)
+
+    def __setitem__(self, key: typing.Any, value: typing.Any) -> None:
+        self._definitions[key] = value
+
+    def __delitem__(self, key: typing.Any) -> None:
+        del self._definitions[key]
+
+
+def set_definitions(field: Field, definitions: SchemaDefinitions) -> None:
+    """
+    Recursively set the definitions that string-referenced `Reference` fields
+    should use.
+    """
+    if isinstance(field, Reference) and field.definitions is None:
+        field.definitions = definitions
+    elif isinstance(field, Array):
+        if field.items is not None:
+            if isinstance(field.items, (tuple, list)):
+                for child in field.items:
+                    set_definitions(child, definitions)
+            else:
+                set_definitions(field.items, definitions)
+    elif isinstance(field, Object):
+        for child in field.properties.values():
+            set_definitions(child, definitions)
 
 
 class SchemaMetaclass(ABCMeta):
@@ -156,3 +194,49 @@ class Schema(Mapping, metaclass=SchemaMetaclass):
         )
         sparse_indicator = " [sparse]" if self.is_sparse else ""
         return f"{class_name}({argument_str}){sparse_indicator}"
+
+
+class Reference(Field):
+    errors = {"null": "May not be null."}
+
+    def __init__(
+        self,
+        to: typing.Union[str, typing.Type[Schema]],
+        definitions: typing.Mapping = None,
+        **kwargs: typing.Any,
+    ) -> None:
+        super().__init__(**kwargs)
+        self.to = to
+        self.definitions = definitions
+        if isinstance(to, str):
+            self._target_string = to
+        else:
+            assert issubclass(to, Schema)
+            self._target = to
+
+    @property
+    def target_string(self) -> str:
+        if not hasattr(self, "_target_string"):
+            self._target_string = self._target.__name__
+        return self._target_string
+
+    @property
+    def target(self) -> typing.Union[Field, typing.Type[Schema]]:
+        if not hasattr(self, "_target"):
+            assert (
+                self.definitions is not None
+            ), "String reference missing 'definitions'."
+            self._target = self.definitions[self.to]
+        return self._target
+
+    def validate(self, value: typing.Any, *, strict: bool = False) -> typing.Any:
+        if value is None and self.allow_null:
+            return None
+        elif value is None:
+            raise self.validation_error("null")
+        return self.target.validate(value, strict=strict)
+
+    def serialize(self, obj: typing.Any) -> typing.Any:
+        if obj is None:
+            return None
+        return dict(obj)
